Slightly restructure management module
