+++ kglib/kgcn/management/grakn/thing.py
